According to BPMN 2.0 schema, all elements can have a documentation
