+++ activiti-bpmn-converter/src/main/java/org/activiti/bpmn/converter/child/DocumentationParser.java
@@ -12,13 +12,11 @@
  */
 package org.activiti.bpmn.converter.child;
 
-import javax.xml.stream.XMLStreamReader;
-
 import org.activiti.bpmn.model.BaseElement;
 import org.activiti.bpmn.model.BpmnModel;
-import org.activiti.bpmn.model.FlowElement;
-import org.activiti.bpmn.model.Process;
 import org.apache.commons.lang3.StringUtils;
+
+import javax.xml.stream.XMLStreamReader;
 
 /**
 
@@ -32,11 +30,7 @@ public class DocumentationParser extends BaseChildElementParser {
   public void parseChildElement(XMLStreamReader xtr, BaseElement parentElement, BpmnModel model) throws Exception {
     String docText = xtr.getElementText();
     if (StringUtils.isNotEmpty(docText)) {
-      if (parentElement instanceof FlowElement) {
-        ((FlowElement) parentElement).setDocumentation(docText.trim());
-      } else if (parentElement instanceof Process) {
-        ((Process) parentElement).setDocumentation(docText.trim());
-      }
+      parentElement.setDocumentation(docText.trim());
     }
   }
 }
